refactor(quiz): modularize results canvas and email gate overlay
